pyproject.toml
[project]
name = "agent-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.8"
dependencies = [
    "agno>=1.7.2",
    "dedent>=0.5",
    "fastapi>=0.116.1",
    "markdown>=3.8.2",
    "openai>=1.96.0",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
    "yfinance>=0.2.65",
]
